src/main/java/com/knkweb/yarnshop/security/config/MyUserDetailService.java
package com.knkweb.yarnshop.security.config;

import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class MyUserDetailService implements UserDetailsService {


    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        System.out.println("fetching userDetail");
        return new myUserDetails(username);
    }
}
